Moved component and messages
