+++ code/src/layouts/subintro/Subintro.test.tsx
@@ -1,0 +1,12 @@
+import { render, screen } from "@testing-library/react";
+import Subintro from "@/layouts/subintro/Subintro";
+import subintroItem from "@/layouts/subintro/components/SubintroItem";
+
+describe("Subintro section", () => {
+  test("renders four subintro elements correctly", () => {
+    render(<Subintro />);
+
+    const subintroItems = screen.getAllByTestId("spa-subintro-item");
+    expect(subintroItems).toHaveLength(subintroItem.length);
+  });
+});
