+++ code/src/layouts/subintro/Subintro.tsx
@@ -1,15 +1,15 @@
 import Box from "@mui/material/Box";
-import subintroData from "@/components/subintro/Subintro.constants";
-import SubintroElement from "@/components/subintroElement/SubintroElement";
+import subintroData from "@/layouts/subintro/Subintro.constants";
+import SubintroItem from "@/layouts/subintro/components/SubintroItem";
 import PageWrapper from "@/layouts/page-wrapper/PageWrapper";
-import "@/components/subintro/subintro.scss";
+import "@/layouts/subintro/subintro.scss";
 
 const Subintro = () => {
   return (
     <PageWrapper>
       <Box className="spa-subintro">
         {subintroData.map((element) => (
-          <SubintroElement
+          <SubintroItem
             icon={element.icon}
             key={element.id}
             content={element.content}
