+++ code/src/layouts/subintro/messages/en.json
@@ -1,0 +1,9 @@
+{
+  "application.header": "Hello, React with Webpack!",
+  "appBanner.header": "Incredible Prices on All Your Favorite Items",
+  "appBanner.paragraphfirst": "Welcome to our store! Here you will find the best gadgets from around the world, and you can choose the perfect product for you.",
+  "subintro.firstParagraph": "Curb-side\npickup",
+  "subintro.secondParagraph": "Free shipping on\norders over $50",
+  "subintro.thirdParagraph": "Low price\nguaranteed",
+  "subintro.fourthParagraph": "Available to\nyou 24/7"
+}
