+++ code/src/layouts/subintro/messages/index.ts
@@ -1,0 +1,4 @@
+import en from "@/layouts/subintro/messages/en.json";
+import uk from "@/layouts/subintro/messages/uk.json";
+
+export default { en, uk };
