+++ code/src/layouts/subintro/messages/uk.json
@@ -1,0 +1,7 @@
+{
+  "appBanner.paragraphfirst": "Ласкаво просимо до нашого магазину! Тут ви знайдете найкращі гаджети з усього світу та зможете обрати найкращий для вас товар.",
+  "subintro.firstParagraph": "Підбір замовлення біля магазину",
+  "subintro.secondParagraph": "Безкоштовна доставка замовлень на\n суму понад $50",
+  "subintro.thirdpPragraph": "Низька ціна\n гарантована",
+  "subintro.fourthParagraph": "Доступний для\n вас 24/7"
+}
